refactor(backend): extract zap endpoint to services/tasks/zaps.ts (U3)
Part of the tasks.ts mega-file split (CQ-H1). Moves getZapsForTask to
a sibling submodule. Drops the now-unused gt + isNotNull drizzle
imports from the parent.

services/tasks.ts: 807 -> 751 LOC. Under the 800-line budget. Plan
complete: 1286 -> 751 (-535 net out of the parent, distributed across
three sibling submodules and a shared internals helper).

Final layout:
- services/tasks.ts            751 LOC  (generation, assignment,
                                         progress, queries + barrel)
- services/tasks/retry.ts      420 LOC  (handleTaskFailure,
                                         reassignStaleTasks, MAX_RETRIES)
- services/tasks/agent-projection.ts  99 LOC  (listTasksByAgent,
                                               projectAgentTaskRows)
- services/tasks/zaps.ts        71 LOC  (getZapsForTask)
- services/tasks/_internals.ts  32 LOC  (jsonSafeBigint helper)

Validation: just check green, 564 backend tests + 30 isolated
tasks.test.ts + 17 agent-heartbeat integration all pass.
